localStorage 이용하여 로그인 유지 구현

+++ src/main/java/com/project/cobell/dto/UserDto.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class UserDto {
-	private int num;
+	private Long num;
 	private String id;
 	private String password;
 	private String nickname;
